+++ crates/services/gas_price_service/src/v1.rs
@@ -1,2 +1,2 @@
 pub mod algorithm;
-pub mod da_source_adapter;
+pub mod da_source_service;
